+++ module/These/src/These/Form/CoEncadrant/RechercherCoEncadrantForm.php
@@ -9,29 +9,42 @@ use UnicaenApp\Form\Element\SearchAndSelect;
 
 class RechercherCoEncadrantForm extends Form {
 
-    private $urlCoEncadrant;
+    private ?string $urlCoEncadrant = null;
+    private ?string $urlEtablissement = null;
 
-    /**
-     * @param string $urlCoEncadrant
-     */
     public function setUrlCoEncadrant(string $urlCoEncadrant): void
     {
         $this->urlCoEncadrant = $urlCoEncadrant;
         $this->get('co-encadrant')->setAutocompleteSource($this->urlCoEncadrant);
     }
+    public function setUrlEtablisssement(string $urlEtablissement): void
+    {
+        $this->urlEtablissement = $urlEtablissement;
+        $this->get('etablissement')->setAutocompleteSource($this->urlEtablissement);
+    }
 
-    public function init()
+    public function init(): void
     {
         /**
          * SearchAndSelect sur les Individus de la structure fictives
          */
-        $coEncadrant = new SearchAndSelect('co-encadrant', ['label' => "Co-encadrant * :"]);
+        $coEncadrant = new SearchAndSelect('co-encadrant', ['label' => "Co-encadrant·e * :"]);
         $coEncadrant
             ->setAutocompleteSource($this->urlCoEncadrant)
             ->setSelectionRequired(true)
             ->setAttributes([
                 'id' => 'co-encadrant',
-                'placeholder' => "Sélectionner un co-encadrant ... ",
+                'placeholder' => "Sélectionner un·e co-encadrant·e ... ",
+            ]);
+        $this->add($coEncadrant);
+
+        $coEncadrant = new SearchAndSelect('etablissement', ['label' => "Établissement (laisser vide si établissement d'inscription) :"]);
+        $coEncadrant
+            ->setAutocompleteSource($this->urlEtablissement)
+            ->setSelectionRequired(true)
+            ->setAttributes([
+                'id' => 'etablissement',
+                'placeholder' => "Sélectionner un établissement ... ",
             ]);
         $this->add($coEncadrant);
 
@@ -52,6 +65,7 @@ class RechercherCoEncadrantForm extends Form {
 
         $this->setInputFilter((new Factory())->createInputFilter([
             'co-encadrant' => [ 'required' => true ],
+            'etablissement' => [ 'required' => false ],
         ]));
     }
 }