+++ module/These/src/These/Service/CoEncadrant/CoEncadrantService.php
@@ -2,6 +2,8 @@
 
 namespace These\Service\CoEncadrant;
 
+use Application\QueryBuilder\DefaultQueryBuilder;
+use Doctrine\ORM\Exception\NotSupported;
 use Doctrine\ORM\NonUniqueResultException;
 use Doctrine\ORM\QueryBuilder;
 use InvalidArgumentException;
@@ -14,76 +16,59 @@ use These\Entity\Db\These;
 use UnicaenApp\Exception\RuntimeException;
 use UnicaenApp\Service\EntityManagerAwareTrait;
 
-class CoEncadrantService {
+class CoEncadrantService
+{
     use EntityManagerAwareTrait;
 
-    /**
-     * @return QueryBuilder
-     */
-    public function createQueryBuilder() {
-        $qb = $this->getEntityManager()->getRepository(Acteur::class)->createQueryBuilder('acteur')
-            ->addSelect('individu')->join('acteur.individu', 'individu')
-            ->addSelect('role')->join('acteur.role', 'role')
-            ->andWhere('role.code = :code')
-            ->setParameter('code', 'B')
-        ;
+    public function createQueryBuilder(): QueryBuilder
+    {
+        try {
+            $qb = $this->getEntityManager()->getRepository(Acteur::class)->createQueryBuilder('acteur')
+                ->addSelect('individu')->join('acteur.individu', 'individu')
+                ->addSelect('role')->join('acteur.role', 'role')
+                ->andWhere('role.code = :code')
+                ->setParameter('code', 'B');
+        } catch (NotSupported $e) {
+            throw new RuntimeException("Un problème est survenu lors de la création du query builder", 0, $e);
+        }
         return $qb;
     }
 
-    /**
-     * @param string $term
-     * @return Acteur[]
-     */
-    public function findByText(string $term)
+    /**@return Acteur[] */
+    public function findByText(string $term): array
     {
         $qb = $this->createQueryBuilder()
             ->andWhere("LOWER(concat(concat(concat(concat(individu.prenom1, ' '),individu.nomUsuel), ' '), individu.prenom1)) like :term")
-            ->setParameter('term', '%'.strtolower($term).'%')
-        ;
+            ->setParameter('term', '%' . strtolower($term) . '%');
         $result = $qb->getQuery()->getResult();
         return $result;
     }
 
-    /**
-     * @param int|null $id
-     * @return Acteur|null
-     */
-    public function getCoEncadrant(?int $id) : ?Acteur
+    public function getCoEncadrant(?int $id): ?Acteur
     {
         if ($id === null) return null;
 
         $qb = $this->createQueryBuilder()
             ->andWhere('acteur.id = :id')
-            ->setParameter('id', $id)
-        ;
+            ->setParameter('id', $id);
         try {
             $result = $qb->getQuery()->getOneOrNullResult();
         } catch (NonUniqueResultException $e) {
-            throw new RuntimeException("Plusieurs Acteur partagent le même id [".$id."].");
+            throw new RuntimeException("Plusieurs Acteur partagent le même id [" . $id . "].", 0, $e);
         }
         return $result;
     }
 
-    /**
-     * @param AbstractActionController $controller
-     * @param string $param
-     * @return ?Acteur
-     */
-    public function getRequestedCoEncadrant(AbstractActionController $controller, string $param = 'co-encadrant') : ?Acteur
+    public function getRequestedCoEncadrant(AbstractActionController $controller, string $param = 'co-encadrant'): ?Acteur
     {
         $id = $controller->params()->fromRoute($param);
         $result = $this->getCoEncadrant($id);
         return $result;
     }
 
-    /**
-     * @param EcoleDoctorale|UniteRecherche $structureConcrete
-     * @param bool $encours
-     * @return array
-     */
     public function findCoEncadrantsByStructureConcrete(StructureConcreteInterface $structureConcrete, bool $encours = true): array
     {
-        /** @var \Application\QueryBuilder\DefaultQueryBuilder $qb */
+        /** @var DefaultQueryBuilder $qb */
         $qb = $this->createQueryBuilder()
             ->addSelect('these')->join('acteur.these', 'these')
             ->andWhere('these.histoDestruction is null')